Changed QuestLogGump in XMLQuestLogGump because of name conflicts with RunUO source from SVN.
In XmlSpawner2.cs fixed the SpawnerBounds, that returned a smaller area than the real one of the spawner, this only affected FOREACH command.

Fixed an old bug in XmlDialog, where an NPC with LockedConversation selected, if it starts a conversation and then waits for input from the user, another user giving speech for that input could obtain, in case of successful eventual checks, the control of the subsequent input, now this is checked, so LockedConversation makes more sense.

Added some extras found in the web, they were not checked for any bugs, so, if you plan on using them, be warned that they may cause server instability, TEST THEM at your own risk!